Moving image pullers to images directory
